fix(appointments): remove dead participant routes, collapse the live two
Four participant-roster pages existed; only two were reachable.

`supportsParticipantManagement()` has always returned true only for
webinar/class, so nothing ever linked to the consultations and
subscriptions pages — while `getParticipantManagementUrl()` kept emitting
hrefs to them. That gate is correct, not a bug: both are 1-on-1, and the
consultation page rendered `participants.length/2 (1-on-1 consultation)`,
i.e. a "roster" of the consultant and the one attendee. Deleted both page
trees and the two dead switch branches rather than wiring up a roster of
two. Their API routes are left alone — they serve other callers.

The remaining webinars/classes pages were ~270 lines each and are now one
`participants/[eventType]/[eventId]` route. Most of the difference was the
entity noun and the API path segment, but not all of it: a class carries
`appointments: Appointment[]` where a webinar carries a single
`appointment | null`, so the response is narrowed on the key rather than
cast through an index signature, and the appointment list is normalised
before the participant flattening. An unknown eventType 404s.

889 -> 290 LOC.

Part of #1021
