+++ app/dashboard/consultant/[consultantId]/(features)/appointments/participants/[eventType]/[eventId]/page.tsx
@@ -1,0 +1,293 @@
+"use client";
+
+/**
+ * Participant roster for a group session — one route for webinars and classes.
+ *
+ * There were four of these: `webinars/[webinarId]`, `classes/[classId]`,
+ * `consultations/[consultationId]` and `subscriptions/[subscriptionId]`.
+ *
+ * The last two were unreachable — `supportsParticipantManagement()` only
+ * returns true for webinar/class, so nothing ever linked to them — and that
+ * gate is right: consultations and subscriptions are 1-on-1, so their
+ * "roster" was a list of two people (the consultation page literally rendered
+ * `participants.length/2 (1-on-1 consultation)`). They're deleted rather than
+ * wired up, along with the `getParticipantManagementUrl` branches that
+ * emitted hrefs to them.
+ *
+ * The remaining two were ~270 lines each and collapse to this one route. Most
+ * of the difference was the entity noun (`webinarEvent`/`webinarPlan` vs
+ * `classEvent`/`classPlan`) and the API path segment, but not all of it: a
+ * class carries `appointments: Appointment[]` while a webinar carries a single
+ * `appointment | null`, so the two are narrowed rather than cast and the
+ * appointment list is normalised before the participant flattening.
+ */
+
+import * as Sentry from "@sentry/nextjs";
+import Link from "next/link";
+import { notFound, useParams } from "next/navigation";
+import { ArrowLeft } from "lucide-react";
+import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
+import { format } from "date-fns";
+
+import { Card, CardContent } from "@/components/ui/card";
+import { Button } from "@/components/ui/button";
+import { Badge } from "@/components/ui/badge";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import {
+  ResponsiveTable,
+  type ResponsiveColumn,
+} from "@/components/ui/responsive-table";
+import { DashboardHeader } from "@/components/dashboard/PageScaffold";
+import type { WaitlistParticipant } from "@/types/participants";
+
+import type { ClassEvent, WebinarEvent } from "../../../types/event";
+
+/** URL segment → API path segment and the noun used in the count line. */
+const EVENT_KINDS = {
+  webinars: { apiSegment: "webinar", countNoun: "registered" },
+  classes: { apiSegment: "class", countNoun: "participants" },
+} as const;
+
+type EventKind = keyof typeof EVENT_KINDS;
+
+/**
+ * The two responses differ in more than naming: a class carries
+ * `appointments: Appointment[]`, a webinar a single `appointment | null`.
+ * Narrowing on the key rather than casting keeps that difference honest.
+ */
+type ParticipantsResponse = { waitlist?: WaitlistParticipant[] } & (
+  | { webinarEvent: WebinarEvent; classEvent?: never }
+  | { classEvent: ClassEvent; webinarEvent?: never }
+);
+
+// Registered-participant rows are flattened from the event's slot users.
+type RegisteredParticipant = { id: string; name?: string; email?: string };
+
+// Waitlist status pill colors — semantic, kept with dark: variants.
+function getWaitlistStatusColor(status: string): string {
+  if (status === "NOTIFIED")
+    return "bg-amber-100 text-amber-800 dark:bg-amber-900/30 dark:text-amber-300";
+  if (status === "EXPIRED")
+    return "bg-red-100 text-red-800 dark:bg-red-900/30 dark:text-red-300";
+  return "bg-muted text-muted-foreground";
+}
+
+const fetchParticipants = async (
+  apiSegment: string,
+  eventId: string,
+): Promise<ParticipantsResponse> => {
+  const response = await fetch(`/api/participants/${apiSegment}/${eventId}`);
+  if (!response.ok) throw new Error("Failed to fetch event data");
+  return response.json();
+};
+
+const removeParticipant = async ({
+  apiSegment,
+  eventId,
+  userId,
+}: {
+  apiSegment: string;
+  eventId: string;
+  userId: string;
+}) => {
+  const response = await fetch(
+    `/api/participants/${apiSegment}/${eventId}?userId=${userId}`,
+    { method: "DELETE" },
+  );
+  if (!response.ok) throw new Error("Failed to remove participant");
+  return response.json();
+};
+
+export default function EventParticipantsPage() {
+  const params = useParams();
+  const queryClient = useQueryClient();
+
+  const eventType = params.eventType as string;
+  const eventId = params.eventId as string;
+  const kind = EVENT_KINDS[eventType as EventKind];
+
+  const { data, isLoading, error } = useQuery({
+    queryKey: ["event-participants", eventType, eventId],
+    queryFn: () => fetchParticipants(kind.apiSegment, eventId),
+    enabled: !!eventId && !!kind,
+  });
+
+  const removeParticipantMutation = useMutation({
+    mutationFn: removeParticipant,
+    onSuccess: () => {
+      queryClient.invalidateQueries({
+        queryKey: ["event-participants", eventType, eventId],
+      });
+    },
+    onError: (err) => {
+      Sentry.captureException(
+        err instanceof Error ? err : new Error(String(err)),
+        { tags: { subsystem: "client" } },
+      );
+      console.error("Error removing participant:", err);
+    },
+  });
+
+  // Only webinars and classes have a roster worth managing. Anything else in
+  // the URL is a hand-edit or a stale link.
+  if (!kind) notFound();
+
+  const handleRemoveParticipant = (userId: string) => {
+    removeParticipantMutation.mutate({
+      apiSegment: kind.apiSegment,
+      eventId,
+      userId,
+    });
+  };
+
+  if (isLoading) return <div>Loading...</div>;
+  if (error) return <div>Error loading event data</div>;
+
+  const event = data?.webinarEvent ?? data?.classEvent;
+  if (!event) return <div>Event not found</div>;
+
+  const waitlist = data?.waitlist ?? [];
+  const plan = "webinarPlan" in event ? event.webinarPlan : event.classPlan;
+
+  // A class has many appointments; a webinar has at most one. Normalise so
+  // the flattening below is identical for both.
+  const appointments =
+    "appointments" in event
+      ? event.appointments
+      : event.appointment
+        ? [event.appointment]
+        : [];
+
+  // Unique participants by user id, flattened out of the event's slots.
+  const participants = Array.from(
+    new Map(
+      appointments
+        .flatMap((appointment) =>
+          (appointment.slotsOfAppointment || []).flatMap(
+            (slot) => slot.user || [],
+          ),
+        )
+        .map((user) => [user.id, user]),
+    ).values(),
+  );
+
+  const registeredColumns: ResponsiveColumn<RegisteredParticipant>[] = [
+    { key: "name", header: "Name", primary: true, cell: (p) => p.name },
+    { key: "email", header: "Email", cell: (p) => p.email },
+    {
+      key: "status",
+      header: "Status",
+      cell: () => (
+        <span className="inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium bg-green-100 text-green-800 dark:bg-green-900/30 dark:text-green-300">
+          Registered
+        </span>
+      ),
+    },
+  ];
+
+  const renderRegisteredActions = (participant: RegisteredParticipant) => (
+    <Button
+      variant="outline"
+      size="sm"
+      onClick={() => handleRemoveParticipant(participant.id)}
+      disabled={removeParticipantMutation.isPending}
+    >
+      {removeParticipantMutation.isPending ? "Removing..." : "Remove"}
+    </Button>
+  );
+
+  const waitlistColumns: ResponsiveColumn<WaitlistParticipant>[] = [
+    {
+      key: "position",
+      header: "Position",
+      cell: (entry) => (
+        <Badge variant="outline" className="font-mono">
+          #{entry.position ?? "-"}
+        </Badge>
+      ),
+    },
+    { key: "name", header: "Name", primary: true, cell: (e) => e.user.name },
+    { key: "email", header: "Email", cell: (e) => e.user.email },
+    {
+      key: "joined",
+      header: "Joined",
+      cell: (e) => format(new Date(e.joinedAt), "MMM d, yyyy h:mm a"),
+    },
+    {
+      key: "status",
+      header: "Status",
+      cell: (e) => (
+        <span
+          className={`inline-flex items-center px-2.5 py-0.5 rounded-full text-xs font-medium ${getWaitlistStatusColor(
+            e.status,
+          )}`}
+        >
+          {e.status}
+        </span>
+      ),
+    },
+  ];
+
+  return (
+    <>
+      <DashboardHeader
+        title={`${plan.title} — Participants`}
+        subtitle={`${participants.length}/${plan.maxParticipants} ${kind.countNoun} · ${waitlist.length} on waitlist`}
+        actions={
+          <Link
+            href={`/dashboard/consultant/${params.consultantId}/appointments`}
+            passHref
+          >
+            <Button variant="outline" size="sm">
+              <ArrowLeft className="mr-2 h-4 w-4" /> Back to Appointments
+            </Button>
+          </Link>
+        }
+      />
+      <Card className="mt-6">
+        <CardContent className="pt-6">
+          <Tabs defaultValue="registered" className="w-full">
+            <TabsList className="mb-4">
+              <TabsTrigger value="registered">Registered</TabsTrigger>
+              <TabsTrigger value="waitlist">
+                Waitlist
+                {waitlist.length > 0 && (
+                  <Badge variant="secondary" className="ml-2 h-5 px-1.5 min-w-5">
+                    {waitlist.length}
+                  </Badge>
+                )}
+              </TabsTrigger>
+            </TabsList>
+
+            <TabsContent value="registered">
+              <ResponsiveTable<RegisteredParticipant>
+                columns={registeredColumns}
+                rows={participants}
+                getRowId={(p) => p.id}
+                rowActions={renderRegisteredActions}
+                empty={
+                  <div className="py-8 text-center text-muted-foreground">
+                    No registered participants yet.
+                  </div>
+                }
+              />
+            </TabsContent>
+
+            <TabsContent value="waitlist">
+              <ResponsiveTable<WaitlistParticipant>
+                columns={waitlistColumns}
+                rows={waitlist}
+                getRowId={(e) => e.id}
+                empty={
+                  <div className="py-8 text-center text-muted-foreground">
+                    Waitlist is empty.
+                  </div>
+                }
+              />
+            </TabsContent>
+          </Tabs>
+        </CardContent>
+      </Card>
+    </>
+  );
+}
