--- app/dashboard/consultant/[consultantId]/(features)/appointments/participants/subscriptions/[subscriptionId]/loading.tsx
@@ -1,5 +1,0 @@
-import { TableSkeleton } from "@/components/dashboard/DashboardSkeletons";
-
-export default function Loading() {
-  return <TableSkeleton />;
-}
